Simplify Chelonia setup in main.js
