+++ frontend/model/notifications/templates.js
@@ -1,10 +1,14 @@
+import type { GIMessage } from '~/shared/domains/chelonia/chelonia.js'
 import type {
   NewProposalType,
   NotificationTemplate
 } from './types.flow.js'
 
-import L from '~/frontend/views/utils/translations.js'
+import sbp from '@sbp/sbp'
 
+import L, { LTags } from '~/frontend/views/utils/translations.js'
+
+const contractName = (contractID) => sbp('state/vuex/state').contracts[contractID]?.type ?? contractID
 // Note: this escaping is not intended as a protection against XSS.
 // It is only done to enable correct rendering of special characters in usernames.
 // To guard against XSS when rendering usernames, use the `v-safe-html` directive.
@@ -12,6 +16,27 @@ const escapeForHtml = (text) => text.replace(/[<>&]/g, '\\$&')
 const strong = (text) => `<strong>${escapeForHtml(text)}</strong>`
 
 export default ({
+  CHELONIA_ERROR (data: { activity: string, error: Error, message: GIMessage }) {
+    const { activity, error, message } = data
+    const contractID = message.contractID()
+    const [opType] = message.op()
+    const { action, meta } = message.decryptedValue()
+    return {
+      body: L("{errName} during {activity} for '{action}' from {b_}{who}{_b} to '{contract}': '{errMsg}'", {
+        ...LTags('b'),
+        errName: error.name,
+        activity,
+        action: action ?? opType,
+        who: meta?.username ?? 'TODO: signing keyID',
+        contract: contractName(contractID),
+        errMsg: error.message ?? '?'
+      }),
+      icon: 'exclamation-triangle',
+      level: 'danger',
+      linkTo: `/app/dashboard?modal=UserSettingsModal&section=application-logs&errorMsg=${encodeURI(error.message)}`,
+      scope: 'app'
+    }
+  },
   GENERAL (data: { contractID: string, message: string }) {
     return {
       body: data.message,
