+++ frontend/views/utils/ui.js
@@ -2,9 +2,10 @@
 
 import sbp from '@sbp/sbp'
 import { OPEN_MODAL, MODAL_RESPONSE } from '@utils/events.js'
+import L, { LError } from '@view-utils/translations.js'
 
 export default (sbp('sbp/selectors/register', {
-  'gi.ui/prompt': function (params: Object): Promise<*> {
+  'gi.ui/prompt' (params: Object): Promise<*> {
     sbp('okTurtles.events/emit', OPEN_MODAL, 'Prompt', null, params)
 
     return new Promise((resolve, reject) => {
@@ -12,5 +13,10 @@ export default (sbp('sbp/selectors/register', {
         resolve(response)
       })
     })
+  },
+  'gi.ui/seriousErrorBanner' (e: Error): void {
+    sbp('okTurtles.data/get', 'BANNER').danger(
+      L('Fatal error: {reportError}', LError(e)), 'exclamation-triangle'
+    )
   }
 }): Object)
